process-data/requirements.txt
python-docx>=1.1.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
pywin32>=306
tqdm>=4.66.0
